requirements.txt
# NeuroResume — Phase 1 + Phase 2 Dependencies

# ── Core ───────────────────────────────────────────────────────────────────────
streamlit>=1.32.0
python-dotenv>=1.0.0

# ── AI / Grok ─────────────────────────────────────────────────────────────────
openai>=1.12.0

# ── Phase 2: Scraping ─────────────────────────────────────────────────────────
playwright>=1.42.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0

# ── Phase 5: Gmail API ─────────────────────────────────────────────────────────
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0

# ── Phase 9: Voice Interface ────────────────────────────────────────────────────
openai>=1.12.0       # Whisper API (optional - only if you have OpenAI key)
# Note: Browser Speech API requires no installation (built into Chrome/Edge/Firefox)

# ── Add-On 9: Chrome Extension API Server ────────────────────────────────────
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
